fix the state selection dropdown in the header
This fix adjusts the z-index of the form in the header, so that it can
be used. It also changes the location that the form sends a user to;
previously it was /ahcbp/pages/<State Name>.html, now it is a path that
begins with /borders, which aligns it with the rest of the site's
structure.